backend/requirements.txt
fastapi==0.100.3
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
google-generativeai==0.15.0
pydantic==1.10.11

